refactor: 로또의 당첨 금액을 저장하는 자료형을 long으로 수정

+++ src/main/java/lotto/service/LottoResultService.java
@@ -24,7 +24,7 @@ public class LottoResultService {
     }
 
     public ReturnRate calculateReturnRate(LottoResult lottoResult, LottoPurchasePrice lottoPurchasePrice) {
-        int totalPrizeMoney = calculateTotalPrizeMoney(lottoResult);
+        long totalPrizeMoney = calculateTotalPrizeMoney(lottoResult);
         return new ReturnRate((double) totalPrizeMoney / lottoPurchasePrice.price() * 100);
     }
 
@@ -45,7 +45,7 @@ public class LottoResultService {
         return lotto.getNumbers().contains(bonusNumber.number());
     }
 
-    private int calculateTotalPrizeMoney(LottoResult lottoResult) {
+    private long calculateTotalPrizeMoney(LottoResult lottoResult) {
         return Arrays.stream(LottoRank.values())
                 .mapToInt(rank -> lottoResult.getWinningCount(rank) * rank.getPrizeMoney())
                 .sum();
